requirements.txt
alembic==1.4.2
appdirs==1.4.4
Babel==2.8.0
beautifulsoup4==4.8.2
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cycler==0.10.0
distlib==0.3.0
filelock==3.0.12
firebase-admin==4.2.0
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
google-api-core==1.17.0
google-api-python-client==1.8.2
google-auth==1.14.2
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
gunicorn==20.0.4
httplib2==0.17.3
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.0.1
Mako==1.1.2
MarkupSafe==1.1.1
matplotlib==3.0.3
msgpack==1.0.0
numpy==1.16.2
pipenv==2018.11.26
protobuf==3.11.3
psycopg2-binary==2.8.5
py-dateutil==2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==0.9.3
pyparsing==2.3.1
python-dateutil==2.6.0
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.12.0
soupsieve==2.0
SQLAlchemy==1.3.16
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.21
virtualenv-clone==0.5.4
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
